Extract AlterResourceGroupStatement to reduce duplicate logic for MySQL and doris
